refactor(frontend): ONE agent-overview body — the composition, not just the frame
The components on this page were already shared; the ARRANGEMENT was not, so /m
assembled its own sections in its own order and was missing Files and Usage
entirely. AgentOverviewBody is the whole body — composer slot, both list cards,
and the rail in one order (what the agent IS, CARRIES, WILL DO, COST) — rendered
by the desktop page and the mobile screen alike.

SessionListCard moves to @agenta/sessions-ui with the three app verbs as props
(mobile's useSessionRowMenu already returns exactly those), replacing mobile's
own AgentOverviewSection. AgentFilesCard moves to @agenta/entity-ui/agent,
de-antd'd so mobile can render it. The OSS wrappers those two replaced are gone.
