+++ web/mobile/src/features/agents/AgentOverviewScreen.tsx
@@ -1,13 +1,7 @@
 import {useMemo} from "react"
 
 import {agentWorkflowsListQueryStateAtom, type Workflow} from "@agenta/entities/workflow"
-import {
-    AgentConfigSummaryCard,
-    agentAvatar,
-    AgentOverviewLayout,
-    NextTriggersSection,
-} from "@agenta/entity-ui/agent"
-import {SessionCardList} from "@agenta/sessions-ui"
+import {AgentOverviewBody, agentAvatar} from "@agenta/entity-ui/agent"
 import {useAtomValue} from "jotai"
 
 import {ContentRail} from "@/components/ContentRail"
@@ -18,8 +12,6 @@ import {useBindProjectContext} from "../context/useBindProjectContext"
 import {AppShell} from "../nav/AppShell"
 import {NavDrawer} from "../nav/NavDrawer"
 import {useSessionRowMenu} from "../sessions/useSessionRowMenu"
-
-import {AgentOverviewSection} from "./AgentOverviewSection"
 
 /**
  * One agent's overview — the mobile face of the desktop agent overview page: this agent's
@@ -56,6 +48,7 @@ export const AgentOverviewScreen = ({
             <PageTitle parts={[name, "Agents"]} />
             <AppShell workspaceId={workspaceId} projectId={projectId}>
                 <ScreenScaffold
+                    fill
                     header={
                         <div className="border-border shrink-0 border-b px-2 pb-3 pt-2">
                             <ContentRail className="flex items-center gap-2 lg:max-w-none">
@@ -75,63 +68,18 @@ export const AgentOverviewScreen = ({
                         </div>
                     }
                 >
-                    {/* The SHARED arrangement the desktop overview uses — activity left, the
-                        agent's own state as a rail. `gap-0` below lg keeps the phone's stacked
-                        rhythm, which the sections already own through their own padding. */}
-                    <ContentRail className="lg:max-w-none">
-                        <AgentOverviewLayout
-                            className="gap-0 lg:gap-6"
-                            main={
-                                <>
-                                    <AgentOverviewSection
-                                        title="Sessions"
-                                        viewAllHref={`${base}/sessions`}
-                                    >
-                                        <div className="px-2">
-                                            <SessionCardList
-                                                withPinned
-                                                agentId={agentId}
-                                                limit={6}
-                                                emptyText="Conversations with this agent will show up here."
-                                                onOpenRow={sessionMenu.open}
-                                                menuFor={sessionMenu.menuFor}
-                                                onMenuSelect={sessionMenu.onMenuSelect}
-                                                alwaysShowPin
-                                            />
-                                        </div>
-                                    </AgentOverviewSection>
-
-                                    <AgentOverviewSection title="Automation runs">
-                                        <div className="px-2">
-                                            <SessionCardList
-                                                origin="trigger"
-                                                agentId={agentId}
-                                                limit={5}
-                                                emptyText="Runs from automations bound to this agent will show up here."
-                                                onOpenRow={sessionMenu.open}
-                                                menuFor={sessionMenu.menuFor}
-                                                onMenuSelect={sessionMenu.onMenuSelect}
-                                                alwaysShowPin
-                                            />
-                                        </div>
-                                    </AgentOverviewSection>
-                                </>
-                            }
-                            rail={
-                                <>
-                                    {/* Scoped to this agent; automation RUNS say what already happened. */}
-                                    <div className="px-2 pt-2 lg:pt-0">
-                                        <NextTriggersSection
-                                            agentId={agentId}
-                                            agentNames={agentNames}
-                                        />
-                                    </div>
-
-                                    <div className="px-2 pb-6 pt-2">
-                                        <AgentConfigSummaryCard appId={agentId} />
-                                    </div>
-                                </>
-                            }
+                    {/* THE shared overview body — the same cards, order and chrome the desktop
+                        page renders. Read-only host: configuration is edited in the desktop
+                        playground, so no `onEditConfig`. */}
+                    <ContentRail className="min-h-0 flex-1 px-2 lg:max-w-none">
+                        <AgentOverviewBody
+                            alwaysShowPin
+                            agentId={agentId}
+                            agentNames={agentNames}
+                            sessionsHref={`${base}/sessions`}
+                            onOpenRow={sessionMenu.open}
+                            menuFor={sessionMenu.menuFor}
+                            onMenuSelect={sessionMenu.onMenuSelect}
                         />
                     </ContentRail>
                 </ScreenScaffold>
