+++ web/packages/agenta-entity-ui/src/agent/index.ts
@@ -11,3 +11,5 @@ export {agentLatestRevisionAtomFamily} from "./state"
 export {AgentCardGrid, type AgentCardGridProps} from "./AgentCardGrid"
 export {AgentRosterGrid, type AgentRosterEntry, type AgentRosterGridProps} from "./AgentRosterGrid"
 export {AgentOverviewLayout, type AgentOverviewLayoutProps} from "./AgentOverviewLayout"
+export {AgentFilesCard} from "./AgentFilesCard"
+export {AgentOverviewBody, type AgentOverviewBodyProps} from "./AgentOverviewBody"
